refactor: move classes to helper package
