+++ synapse-client-rest/src/main/java/io/americanexpress/synapse/client/rest/client/BaseRestClient.java
@@ -14,6 +14,7 @@
 package io.americanexpress.synapse.client.rest.client;
 
 import io.americanexpress.synapse.client.rest.factory.BaseClientHttpHeadersFactory;
+import io.americanexpress.synapse.client.rest.helper.UrlBuilder;
 import io.americanexpress.synapse.client.rest.model.BaseClientRequest;
 import io.americanexpress.synapse.client.rest.model.BaseClientResponse;
 import io.americanexpress.synapse.client.rest.model.ClientHeaders;
