+++ synapse-client-rest/src/main/java/io/americanexpress/synapse/client/rest/helper/ClientLoggingCustomizer.java
@@ -11,7 +11,7 @@
  * or implied. See the License for the specific language governing permissions and limitations under
  * the License.
  */
-package io.americanexpress.synapse.client.rest.interceptor;
+package io.americanexpress.synapse.client.rest.helper;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.web.client.RestTemplateCustomizer;
