+++ synapse-client-rest/src/main/java/io/americanexpress/synapse/client/rest/helper/UrlBuilder.java
@@ -1,4 +1,4 @@
-package io.americanexpress.synapse.client.rest.client;
+package io.americanexpress.synapse.client.rest.helper;
 
 import java.util.List;
 import java.util.Optional;
@@ -11,7 +11,7 @@ import io.americanexpress.synapse.client.rest.model.QueryParameter;
  * {@code UrlBuilder} class builds a new URL that includes path variables and/or query parameters.
  * @author Paolo Claudio
  */
-final class UrlBuilder {
+public final class UrlBuilder {
 	
 	/**
 	 * Default constructor creates a new instance of PathVariableUriCreator with default values.
@@ -28,7 +28,7 @@ final class UrlBuilder {
 	 * @param pathVariables path variables
 	 * @return the new URL if any path variables and/or query parameters are present; otherwise the same url will be returned
 	 */
-	static String build(String url, List<QueryParameter> queryParameters, String... pathVariables) {
+	public static String build(String url, List<QueryParameter> queryParameters, String... pathVariables) {
 		return Optional.ofNullable(url)
 			.filter(StringUtils::isNotBlank)
 			.map(uniformResourceLocator -> new StringBuilder(uniformResourceLocator)
