+++ synapse-client-rest/src/main/java/io/americanexpress/synapse/client/rest/model/ClientHeaderKey.java
@@ -11,7 +11,7 @@
  * or implied. See the License for the specific language governing permissions and limitations under
  * the License.
  */
-package io.americanexpress.synapse.client.rest.client;
+package io.americanexpress.synapse.client.rest.model;
 
 /**
  * <code>ClientHeaderKey</code> enum will house all the header keys.
